Module Owl_regression

Regression Module This module implements a set of regression functions. S module provides supports for single precision float numbers whilst D module provides supports for double precision float numbers.

module Make_Embedded (A : Owl_types_ndarray_algodiff.Sig) : sig ... end

module S : sig ... end

module D : sig ... end