owl index

Library owl

This library exposes the following toplevel modules:

Owl

Owl_algodiff

Owl_algodiff_primal_ops

Owl_cblas

Owl_cblas_basic CBLAS interface: high-level interface between Owl and level-1, level-2, level-3 BLAS functions.

Owl_cblas_generated auto-generated cblas interface file, timestamp:1582875912

Owl_cluster

Owl_core_types

Owl_dataset Dataset: easy access to various datasets

Owl_dense Dense data structures: matrix & ndarray

Owl_dense_matrix Matrix: module aliases

Owl_dense_matrix_c Complex dense matrix module: this module supports operations on dense matrices of complex numbers. The complex number has a record type of {re = float; im = float}.

Owl_dense_matrix_d

Owl_dense_matrix_generic Matrix module: including creation, manipulation, and various vectorised mathematical operations.

Owl_dense_matrix_intf

Owl_dense_matrix_s

Owl_dense_matrix_z Complex dense matrix module: this module supports operations on dense matrices of complex numbers. The complex number has a record type of {re = float; im = float}.

Owl_dense_ndarray Ndarray: module aliases

Owl_dense_ndarray_a

Owl_dense_ndarray_c

Owl_dense_ndarray_d

Owl_dense_ndarray_generic N-dimensional array module: including creation, manipulation, and various vectorised mathematical operations.

Owl_dense_ndarray_intf

Owl_dense_ndarray_s

Owl_dense_ndarray_z

Owl_distribution Functor to generate distribution module

Owl_distribution_common Interface to the C implementation of rvs, pdf, and cdf functions.

Owl_distribution_generic

Owl_fft

Owl_fft_d

Owl_fft_generic Fast Fourier Transform

Owl_fft_s

Owl_fftpack

Owl_lapacke

Owl_lapacke_generated LAPACKE interface: low-level interface to the LAPACKE functions

Owl_linalg Linear algebra: module aliases

Owl_linalg_c

Owl_linalg_d

Owl_linalg_generic Linear algebra module including high-level functions to solve linear systems, factorisation, and etc.

Owl_linalg_intf

Owl_linalg_s

Owl_linalg_z

Owl_maths Maths: fundamental and advanced mathematical functions.

Owl_maths_special

Owl_matrix

Owl_matrix_check

Owl_matrix_swap

Owl_ndarray Tensor operations implementation in Owl's Ndarray

Owl_ndarray_contract

Owl_ndarray_conv

Owl_ndarray_fma

Owl_ndarray_maths

Owl_ndarray_pool

Owl_ndarray_repeat

Owl_ndarray_slide

Owl_ndarray_sort

Owl_ndarray_transpose

Owl_ndarray_upsampling

Owl_ndarray_utils

Owl_neural

Owl_neural_parallel Neural network: interface of parallel engine

Owl_nlp NLP: module aliases

Owl_nlp_corpus NLP: Corpus module

Owl_nlp_lda NLP: LDA module

Owl_nlp_similarity

Owl_nlp_tfidf NLP: TFIDF module

Owl_nlp_utils

Owl_nlp_vocabulary NLP: Vocabulary module

Owl_optimise

Owl_regression Regression Module This module implements a set of regression functions. S module provides supports for single precision float numbers whilst D module provides supports for double precision float numbers.

Owl_regression_generic

Owl_regression_generic_sig

Owl_signal Signal: Fundamental Signal Processing functions.

Owl_slicing

Owl_slicing_basic

Owl_slicing_fancy

Owl_stats Statistics: random number generators, PDF and CDF functions, and hypothesis tests. The module also includes some basic statistical functions such as mean, variance, skew, and etc.

Owl_stats_dist

Owl_stats_extend

Owl_stats_prng

Owl_stats_sampler