owl-base index

Library owl-base

This library exposes the following toplevel modules:

Owl_algodiff_check

Owl_algodiff_core

Owl_algodiff_core_sig

Owl_algodiff_generic

Owl_algodiff_generic_sig

Owl_algodiff_graph_convert

Owl_algodiff_graph_convert_sig

Owl_algodiff_ops

Owl_algodiff_ops_builder

Owl_algodiff_ops_builder_sig

Owl_algodiff_ops_sig

Owl_algodiff_reverse

Owl_algodiff_types

Owl_algodiff_types_sig

Owl_base

Owl_base_algodiff_primal_ops

Owl_base_complex

Owl_base_dense_common

Owl_base_dense_matrix_c

Owl_base_dense_matrix_d

Owl_base_dense_matrix_generic Matrix module: including creation, manipulation, and various vectorised mathematical operations.

Owl_base_dense_matrix_intf

Owl_base_dense_matrix_s

Owl_base_dense_matrix_z

Owl_base_dense_ndarray Ndarray: module aliases

Owl_base_dense_ndarray_c

Owl_base_dense_ndarray_d

Owl_base_dense_ndarray_generic N-dimensional array module: including creation, manipulation, and various vectorised mathematical operations.

Owl_base_dense_ndarray_intf

Owl_base_dense_ndarray_s

Owl_base_dense_ndarray_z

Owl_base_linalg_c

Owl_base_linalg_d

Owl_base_linalg_generic

Owl_base_linalg_intf

Owl_base_linalg_s

Owl_base_linalg_z

Owl_base_maths Maths: fundamental and advanced mathematical functions.

Owl_base_slicing

Owl_base_stats Statistics: random number generators, PDF and CDF functions, and hypothesis tests. The module also includes some basic statistical functions such as mean, variance, skew, and etc.

Owl_base_stats_dist_bernoulli

Owl_base_stats_dist_cauchy

Owl_base_stats_dist_exponential

Owl_base_stats_dist_gamma

Owl_base_stats_dist_gaussian

Owl_base_stats_dist_gumbel1

Owl_base_stats_dist_gumbel2

Owl_base_stats_dist_uniform

Owl_base_stats_prng

Owl_computation

Owl_computation_cpu_device

Owl_computation_cpu_engine

Owl_computation_cpu_eval

Owl_computation_cpu_init

Owl_computation_engine This functor takes a device as its input, then it generates the computation graph module without flattening the module hierarchy.

Owl_computation_engine_sig

Owl_computation_graph

Owl_computation_graph_sig

Owl_computation_operator

Owl_computation_operator_sig

Owl_computation_optimiser

Owl_computation_optimiser_sig

Owl_computation_shape

Owl_computation_shape_sig

Owl_computation_symbol

Owl_computation_symbol_sig

Owl_computation_type

Owl_computation_type_sig

Owl_const Metric system: CGS, MKS, SI, and physical constants.

Owl_countmin_sketch

Owl_countmin_sketch_sig

Owl_countmin_table

Owl_dataframe

Owl_exception

Owl_graph Graph module supports basic operations on DAG.

Owl_heavyhitters_sketch

Owl_heavyhitters_sketch_sig

Owl_io

Owl_lazy

Owl_log Log module provides logging functionality.

Owl_maths_interpolate Interpolation and Extrapolation

Owl_maths_quadrature Numerical Integration

Owl_maths_root Root finding algorithms for nonlinear functions

Owl_neural_compiler

Owl_neural_generic Functor to create neural networks of different precision.

Owl_neural_graph Neural network: Graphical neural network

Owl_neural_graph_sig

Owl_neural_neuron Neural network: Neuron definitions

Owl_neural_neuron_sig

Owl_numdiff_generic

Owl_numdiff_generic_sig Numdiff: numerical differentiation module

Owl_operator

Owl_optimise_generic Optimisation engine

Owl_optimise_generic_sig

Owl_pretty Pretty print the n-dimensional array

Owl_types This module defines the types shared by various sub-libraries in Owl. Note that they just wrappers, to find the exact module signature, please refer to the definition in the corresponding module.

Owl_types_common

Owl_types_computation_device

Owl_types_computation_engine

Owl_types_maths_basic

Owl_types_ndarray_algodiff

Owl_types_ndarray_basic

Owl_types_ndarray_compare

Owl_types_ndarray_eltcmp

Owl_types_ndarray_mutable

Owl_types_ndarray_numdiff

Owl_types_operator Operator definitions such as add, sub, mul, and div. This signature defines the functions need to be implemented.

Owl_types_stats_basic

Owl_types_stats_dist

Owl_utils Helper functions used in the library

Owl_utils_array

Owl_utils_heap

Owl_utils_infer_shape

Owl_utils_multimap

Owl_utils_ndarray

Owl_utils_stack

Owl_view View module This module is used to create views atop of an ndarray. The view creation is very light-weighted and avoids copying actual data. You can further create views atop of existing views using slicing functions.